clm3_5_06
Move pftintdat09_clm3_5_05 to the trunk.

+++ test/system/TER.sh
@@ -1,12 +1,12 @@
 #!/bin/sh 
 #
 
-if [ $# -ne 3 ]; then
+if [ $# -ne 7 ]; then
     echo "TER.sh: incorrect number of input arguments" 
     exit 1
 fi
 
-test_name=TER.$1.$2.$3
+test_name=TER.$1.$2.$3.$4.$5.$6.$7
 
 if [ -f ${CLM_TESTDIR}/${test_name}/TestStatus ]; then
     if grep -c PASS ${CLM_TESTDIR}/${test_name}/TestStatus > /dev/null; then
@@ -41,9 +41,9 @@ if [ $? -ne 0 ]; then
 fi
 cd ${rundir}
 
-initial_length=${3%+*}
-restart_length=${3#*+}
-if [ ${initial_length} = $3 ] || [ ${restart_length} = $3 ]; then
+initial_length=${6%+*}
+restart_length=${6#*+}
+if [ ${initial_length} = $6 ] || [ ${restart_length} = $6 ]; then
     echo "TER.sh: error processing input argument for run lengths" 
     echo "FAIL.job${JOBID}" > TestStatus
     exit 4
@@ -52,7 +52,7 @@ fi
 full_length=`expr $initial_length + $restart_length`
 
 echo "TER.sh: calling TSM.sh for smoke test of full length ${full_length}" 
-${CLM_SCRIPTDIR}/TSM.sh $1 $2 $full_length
+${CLM_SCRIPTDIR}/TSM.sh $1 $2 $3 $4 $5 $full_length $7
 rc=$?
 if [ $rc -ne 0 ]; then
     echo "TER.sh: error from TSM.sh= $rc" 
@@ -61,7 +61,7 @@ if [ $rc -ne 0 ]; then
 fi
 
 echo "TER.sh: calling TSM.sh for smoke test of initial length ${initial_length}" 
-${CLM_SCRIPTDIR}/TSM.sh $1 $2 ${initial_length}
+${CLM_SCRIPTDIR}/TSM.sh $1 $2 $3 $4 $5 $initial_length $7
 rc=$?
 if [ $rc -ne 0 ]; then
     echo "TER.sh: error from TSM.sh= $rc" 
@@ -69,64 +69,20 @@ if [ $rc -ne 0 ]; then
     exit 6
 fi
 
-
-cp ${CLM_TESTDIR}/TSM.$1.$2.${initial_length}/*clm* ${rundir}/.
-cp ${CLM_TESTDIR}/TSM.$1.$2.${initial_length}/rpointer* ${rundir}/.
-
-echo "TER.sh: restarting offline; output in ${CLM_TESTDIR}/${test_name}/test.log" 
-echo "TER.sh: call to build-namelist:"
-
-run_length=${restart_length}
-export run_length
-
-restart_type=1
-export restart_type
-
-clm_nrevsn=' '
-export $clm_revsn
-
-${CLM_SCRIPTDIR}/nl_files/$2 
+echo "TER.sh: calling TSM.sh for smoke test of restart length ${restart_length}" 
+${CLM_SCRIPTDIR}/TSM.sh $1 $2 $3 $4 $5 ${restart_length} \
+                         continue+$1.$2.$3.$4.$5.${initial_length}.$7
 rc=$?
-if [ $rc -eq 0 ]; then
-    echo "TER.sh: build-namelist was successful"
-    cat namelist 
-else
-    echo "TER.sh: error building namelist, error= $rc"
-    echo "TER.sh: see ${CLM_TESTDIR}/${test_name}/test.log for details"
+if [ $rc -ne 0 ]; then
+    echo "TER.sh: error from TSM.sh= $rc" 
     echo "FAIL.job${JOBID}" > TestStatus
     exit 7
 fi
 
-echo "TER.sh calling CLM_runcmnd.sh to build run command"
-## modify the # of tasks/threads for restart
-env CLM_THREADS=${CLM_RESTART_THREADS} CLM_TASKS=${CLM_RESTART_TASKS} \
-    ${CLM_SCRIPTDIR}/CLM_runcmnd.sh $1
-rc=$?
-if [ $rc -eq 0 ] && [ -f clm_run_command.txt ]; then
-    read cmnd < clm_run_command.txt
-    echo "TER.sh: clm run command:"
-    echo "        $cmnd ${CLM_TESTDIR}/TCB.$1/clm"
-    rm clm_run_command.txt
-else
-    echo "TER.sh: error building run command; error from CLM_runcmnd.sh= $rc"
-    echo "FAIL.job${JOBID}" > TestStatus
-    exit 8
-fi
-
-${cmnd} ${CLM_TESTDIR}/TCB.$1/clm >> test.log 2>&1
-rc=$?
-if [ $rc -eq 0 ] && grep -c "TERMINATING CLM MODEL" test.log > /dev/null; then
-    echo "TER.sh: restart of offline clm completed successfully"
-else
-    echo "TER.sh: error on restart run of offline clm, error= $rc" 
-    echo "TER.sh: see ${CLM_TESTDIR}/${test_name}/test.log for details"
-    echo "FAIL.job${JOBID}" > TestStatus
-    exit 9
-fi
-
+mv ${CLM_TESTDIR}/TSM.$1.$2.$3.$4.$5.${restart_length}.continue/*.clm?.h*.nc .
 echo "TER.sh: starting b4b comparisons " 
-files_to_compare=`ls *.clm*h*.nc`
-if [ -z "${files_to_compare}" ]; then
+files_to_compare=`ls *.clm?.h*.nc`
+if [ -z " ${files_to_compare}" ]  && [ "$debug" != "YES" ]; then
     echo "TER.sh: error locating files to compare"
     echo "FAIL.job${JOBID}" > TestStatus
     exit 10
@@ -135,9 +91,17 @@ fi
 all_comparisons_good="TRUE"
 for compare_file in ${files_to_compare}; do
 
+    if [ ! -f $compare_file ]; then
+       echo "TER.sh: trouble finding file $compare_file"
+       exit 11
+    fi
+    if [ ! -f ${CLM_TESTDIR}/TSM.$1.$2.$3.$4.$5.${full_length}.$7/$compare_file ]; then
+       echo "TER.sh: trouble finding file ${CLM_TESTDIR}/TSM.$1.$2.$3.$4.$5.${full_length}.$7/$compare_file"
+       exit 12
+    fi
     ${CLM_SCRIPTDIR}/CLM_compare.sh \
 	${compare_file} \
-	${CLM_TESTDIR}/TSM.$1.$2.${full_length}/${compare_file}
+	${CLM_TESTDIR}/TSM.$1.$2.$3.$4.$5.${full_length}.$7/${compare_file}
     rc=$?
     mv cprnc.out cprnc.${compare_file}.out
     if [ $rc -eq 0 ]; then
@@ -159,7 +123,7 @@ if [ ${all_comparisons_good} = "TRUE" ]; then
 else
     echo "TER.sh: at least one file comparison did not pass" 
     echo "FAIL.job${JOBID}" > TestStatus
-    exit 11
+    exit 13
 fi
 
 exit 0
